+++ rotors_gazebo/src/quaternion_to_rpy.cpp
@@ -19,7 +19,7 @@
 
 #include "rotors_gazebo/transform_datatypes.h"
 #include "rotors_gazebo/Matrix3x3.h"
-#include "rotors_gazebo/Quaternion.h" 
+#include "rotors_gazebo/Quaternion.h"
 
 #include <math.h> 
 #include <ros/ros.h>